Bump gem versions

+++ example/rails_app/config/unicorn.rb
@@ -1,4 +1,4 @@
-worker_processes Integer(4)
+worker_processes Integer(1)
 timeout 15
 preload_app true
 pid File.join('tmp', 'pids', 'unicorn.pid')
